backend/requirements.txt
scikit-learn==1.5.2
numpy==1.26.4
joblib==1.4.2
python-telegram-bot==20.7
python-dotenv==1.0.0
stellar-sdk==8.7.0 